hazelcast-tpc-engine/src/test/java/com/hazelcast/internal/tpc/iouring/SocketAddressFactoryTest.java
/*
 * Copyright (c) 2008-2022, Hazelcast, Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.hazelcast.internal.tpc.iouring;

import org.junit.Test;

import java.net.InetSocketAddress;
import java.net.UnknownHostException;
import java.nio.ByteBuffer;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class SocketAddressFactoryTest {

    @Test
    public void test() throws UnknownHostException {
        ByteBuffer byteBuffer = ByteBuffer.allocate(4);
        byteBuffer.put((byte) 127);
        byteBuffer.put((byte) 0);
        byteBuffer.put((byte) 0);
        byteBuffer.put((byte) 1);
        byteBuffer.flip();
        int ip = byteBuffer.getInt();

        InetSocketAddress socketAddress = SocketAddressFactory.createIPv4Address(ip, 4000);

        assertNotNull(socketAddress);
        assertEquals(4000, socketAddress.getPort());
        assertEquals("127.0.0.1", socketAddress.getHostString());
    }
}
